2. Тесты для идеальный вес [#320600]

+++ src/ru/job4j/calculator/Fit.java
@@ -1,19 +1,19 @@
 package ru.job4j.calculator;
 
 public class Fit {
-    public static double manWeight(short heightMan) {
+    public static double manWeight(double heightMan) {
         double rls = (heightMan - 100) * 1.15;
         return rls;
     }
 
-    public static double womanWeight(short heightWoman) {
+    public static double womanWeight(double heightWoman) {
         double rls = (heightWoman - 110) * 1.15;
         return rls;
     }
 
     public static void main(String[] args) {
-        short heightMan = 170;
-        short heightWoman = 165;
+        double heightMan = 170;
+        double heightWoman = 165;
         double man = Fit.manWeight(heightMan);
         double woman = Fit.womanWeight(heightWoman);
         System.out.println("Man 170 is " + man);
